@@ -1,24 +1,27 @@
 name: bmi_calculator
 description: A new Flutter project.
 
-publish_to: "none" 
+publish_to: "none"
+
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=3.0.2 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
-  font_awesome_flutter: ^10.1.0
-  provider: ^6.0.2
-  shared_preferences: ^2.0.13
+
+  cupertino_icons: ^1.0.5
+  provider: ^6.0.5
+  font_awesome_flutter: ^10.4.0
+  shared_preferences: ^2.1.1
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
+
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
@@ -28,4 +31,3 @@ flutter:
         - asset: assets/fonts/Oxygen-Regular.ttf
         - asset: assets/fonts/Oxygen-Bold.ttf
           weight: 700
-          
